build: add special handling for executable targets on Windows
The Windows executable targets for GUI applications should be created as
targeting the Windows subsystem.  The compiler by default constructs the
entry point as `main`, which classifies it as a console application. The
launch of the application will then display a console Window. Hiding the
window manually upon launch is possible, but will result in flicker, and
more importantly is not the right approach.  For now, explicitly specify
the entry point name via unsafe flags.